Extracted Token enumerable into seperate class TokenContainer
